fix: 게임 참여 기록 조회시 id를 int로 찾는 것을 Long으로 통일

+++ src/main/java/io/cavia/trader/module/member/repository/GameParticipationMapper.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Mapper
 public interface GameParticipationMapper {
 
-    List<GameParticipationDto> findByMemberId(int memberId);
+    List<GameParticipationDto> findByMemberId(Long memberId);
 }
