+++ src/main/java/io/cavia/trader/module/member/repository/GameParticipationRepository.java
@@ -5,5 +5,5 @@ import io.cavia.trader.module.member.dto.GameParticipationDto;
 import java.util.List;
 
 public interface GameParticipationRepository {
-    List<GameParticipationDto> findByMemberId(int memberId);
+    List<GameParticipationDto> findByMemberId(Long memberId);
 }
